+++ api/app/wildfire_one/schema_parsers.py
@@ -1,11 +1,16 @@
 """ Parsers that extract fields from WFWX API responses and build ours"""
 
+import math
+import logging
 from datetime import datetime, timezone
+from app.db.models.observations import HourlyActual
 from app.schemas.stations import WeatherStation
 from app.utils.dewpoint import compute_dewpoint
 from app.data.ecodivision_seasons import EcodivisionSeasons
 from app.schemas.observations import WeatherReading
 from app.schemas.hfi_calc import StationDaily
+
+logger = logging.getLogger(__name__)
 
 
 def _parse_station(station) -> WeatherStation:
@@ -70,3 +75,49 @@ def _parse_daily(raw_daily, station_code) -> StationDaily:
         observation_valid=raw_daily.get('observationValidInd', None),
         observation_valid_comment=raw_daily.get('observationValidComment', None)
     )
+
+
+def parse_hourly_actual(station_code: int, hourly_reading: WeatherReading):
+    """ Maps WeatherReading to HourlyActual """
+    temp_valid = hourly_reading.temperature is not None
+    rh_valid = hourly_reading.relative_humidity is not None and validate_metric(
+        hourly_reading.relative_humidity, 0, 100)
+    wdir_valid = hourly_reading.wind_direction is not None and validate_metric(
+        hourly_reading.wind_direction, 0, 360)
+    wspeed_valid = hourly_reading.wind_speed is not None and validate_metric(
+        hourly_reading.wind_speed, 0, math.inf)
+    precip_valid = hourly_reading.precipitation is not None and validate_metric(
+        hourly_reading.precipitation, 0, math.inf)
+
+    is_valid_wfwx = hourly_reading.observation_valid
+    if is_valid_wfwx is False:
+        logger.warning("Invalid hourly received from WF1 API for station code %s at time %s: %s",
+                       station_code,
+                       hourly_reading.datetime.strftime("%b %d %Y %H:%M:%S"),
+                       hourly_reading.observation_valid_comment)
+
+    is_valid = temp_valid and rh_valid and wdir_valid and wspeed_valid and precip_valid and is_valid_wfwx
+
+    return None if (is_valid is False) else HourlyActual(
+        station_code=station_code,
+        weather_date=hourly_reading.datetime,
+        temp_valid=temp_valid,
+        temperature=hourly_reading.temperature,
+        rh_valid=rh_valid,
+        relative_humidity=hourly_reading.relative_humidity,
+        wspeed_valid=wspeed_valid,
+        wind_speed=hourly_reading.wind_speed,
+        wdir_valid=wdir_valid,
+        wind_direction=hourly_reading.wind_direction,
+        precip_valid=precip_valid,
+        precipitation=hourly_reading.precipitation,
+        dewpoint=hourly_reading.dewpoint,
+        ffmc=hourly_reading.ffmc,
+        isi=hourly_reading.isi,
+        fwi=hourly_reading.fwi,
+    )
+
+
+def validate_metric(value, low, high):
+    """ Validate metric with it's range of accepted values """
+    return low <= value <= high
